requirements.txt
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.10.0
